put stam featcher codes in folder
